check todo in ScoreStat if your looking for things todo

+++ src/java/com/thedarkera/api/systems/scoring/ScoreStat.java
@@ -2,27 +2,34 @@ package com.thedarkera.api.systems.scoring;
 
 public class ScoreStat {
 
-	private static int totalScore;
-	private static int maxScore = 25; // Every time a achievement is added, add to this.
-	private static int addedScore;
-	private static int scorePercentage;
+	private static double totalScore;
+	private static double maxScore = 25.00; // Every time a achievement is added, add to this.
+	private static double addedScore;
+	private static double scorePercentage;
 
-	public static void addScore(int addScore) {
+	public static void addScore(double addScore) {
 		totalScore = totalScore + addScore;
 		addedScore = addScore;
 		scorePercentage = totalScore / maxScore * 100;
 		System.out.println("ScoreSystem: " + "Current Amount of added score is - " + addedScore);
 	}
-	
-	public static int getScore() {
+
+	public static double getScore() {
 		return totalScore;
 	}
 
-	public static int getMaxScore() {
+	public static double getMaxScore() {
 		return maxScore;
 	}
-	
-	public static int getScorePercentage() {
+
+	public static double getScorePercentage() {
 		return scorePercentage;
 	}
+
+	/**
+	 * Achievement Score TODO: 
+	 * All doubles need to be 00.00 not 00.000000
+	 * Change to if achievement = achieved addScore(1)
+	 * Improve Score display
+	 */
 }
